requirements.txt
langchain
langchain-core>=0.2.26
python-dotenv
langchain_openai
langchain_groq
langchain_community
faiss-cpu
rank_bm25
langchain_cohere
langgraph
streamlit
pyPDF2
pandas
langchain-google-genai